+++ src/common/crypto_ed25519.c
@@ -12,6 +12,7 @@
 
 #include "crypto_curve25519.h"
 #include "crypto_ed25519.h"
+#include "crypto_format.h"
 #include "torlog.h"
 #include "util.h"
