+++ src/or/keypin.c
@@ -6,6 +6,7 @@
 #include "orconfig.h"
 #include "compat.h"
 #include "crypto.h"
+#include "crypto_format.h"
 #include "di_ops.h"
 #include "ht.h"
 #include "keypin.h"
@@ -13,6 +14,7 @@
 #include "torint.h"
 #include "torlog.h"
 #include "util.h"
+#include "util_format.h"
 
 #ifdef HAVE_UNISTD_H
 #include <unistd.h>
